Update template imports to avoid warnings
These changes resolve warnings on unused imports in the test file. In one case,
the import of console2, we suppress the warning because it's convenient to leave
console2 imported in test files even if not actively used, so it doesn't first have
to be imported by the dev when debugging. Second, we change the deploy script to
return the Counter rather than store it locally. This change resolves the import issue
in an opinionated way: it's better for the script to be stateless anyway.